fix tính năng xóa movie

+++ ViewModels/MovieManagementVM/Feature/DeleteMovie.cs
@@ -22,24 +22,35 @@ namespace CineMajestic.ViewModels.MovieManagementVM
         }
         private void deleteMovie(object obj)
         {
-            MovieDA movieDA = new MovieDA();
-            //trước khi xóa thì phải đưa thằng khóa ngoại tham chiếu tới nó là null
-            BillAddMovieDA billAddMovieDA = new BillAddMovieDA();
-            billAddMovieDA.updateMovie_IdNull((obj as MovieDTO).Id);
-
-            YesNoMessageBox mb = new YesNoMessageBox("Thông báo", "Bạn có muốn xóa phim này?");
-            mb.ShowDialog();
-            if (mb.DialogResult == false)
-                return;
-            else
+            if (obj != null)
             {
-                //tiến hành xóa movie
-                movieDA.deleteMovie(obj as MovieDTO);
-                mb.Close();
-                YesMessageBox msb = new YesMessageBox("Thông báo", "Xóa thành công");
-                msb.ShowDialog();
-                msb.Close();
-                loadData();
+                //kiểm tra phim đang chiếu thì không được xóa
+                ShowTimeDA showTimeDA = new ShowTimeDA();
+                if (showTimeDA.checkShowtimeByMovie((obj as MovieDTO).Id))
+                {
+                    MessageBox.Show("Phim đang có trong ít nhất 1 suất chiếu,bạn không được phép xóa!");
+                    return;
+                }
+
+                MovieDA movieDA = new MovieDA();
+                //trước khi xóa thì phải đưa thằng khóa ngoại tham chiếu tới nó là null
+                BillAddMovieDA billAddMovieDA = new BillAddMovieDA();
+                billAddMovieDA.updateMovie_IdNull((obj as MovieDTO).Id);
+
+                YesNoMessageBox mb = new YesNoMessageBox("Thông báo", "Bạn có muốn xóa phim này?");
+                mb.ShowDialog();
+                if (mb.DialogResult == false)
+                    return;
+                else
+                {
+                    //tiến hành xóa movie
+                    movieDA.deleteMovie(obj as MovieDTO);
+                    mb.Close();
+                    YesMessageBox msb = new YesMessageBox("Thông báo", "Xóa thành công");
+                    msb.ShowDialog();
+                    msb.Close();
+                    loadData();
+                }
             }
            
         }
